+++ src/Admin/AdminRosters/AdminRosterDetail.jsx
@@ -3,6 +3,7 @@ import appClient from "../../appClient";
 import _ from "lodash";
 import Spinner from "../../Spinner/Spinner";
 import AdminRoster from "./AdminRoster";
+import { Link } from "react-router-dom";
 
 class AdminRosterDetail extends Component {
   state = {
@@ -37,8 +38,6 @@ class AdminRosterDetail extends Component {
         <p className="lead">
           {this.state.camp.name} {_.capitalize(this.state.camp.type)} Roster
         </p>
-        <AdminRoster camp={this.state.camp} />
-        <hr />
         <ul>
           <li>
             <a
@@ -70,7 +69,13 @@ class AdminRosterDetail extends Component {
               Download Swimming Report (.csv)
             </a>
           </li>
+          <li>
+            <Link to={`/admin/rosters/${this.state.camp._id}/swimming`}>
+              Update Swimming Ability
+            </Link>
+          </li>
         </ul>
+        <AdminRoster camp={this.state.camp} />
       </div>
     );
   }